IT4, IT4-H19

+++ src/app/repositorios/firebase/route-firebase.service.ts
@@ -10,6 +10,7 @@ import { Place } from '../../modelos/place';
 import { NotExistingObjectException } from '../../excepciones/notExistingObjectException';
 import { AuthStateService } from '../../utils/auth-state.service';
 import { getAuth } from 'firebase/auth';
+import { ServerNotOperativeException } from '../../excepciones/server-not-operative-exception';
 import { PlaceNotFoundException } from '../../excepciones/place-not-found-exception';
 
 @Injectable({
@@ -106,9 +107,13 @@ export class RouteFirebaseService implements RouteRepository{
     return newRoute;
   }
   
-  async deleteRoute(nombre: string): Promise<void> {
+  async deleteRoute(nombre: string): Promise<boolean> {
+    if (this._authState.currentUser == null) {
+      throw new ServerNotOperativeException();
+    }
     const uid = this._authState.currentUser?.uid;
     await this._firestore.deleteRoute(`ruta/${uid}/listaRutasInterés`, nombre);
+    return true;
   }
 
   async getRoutes(): Promise<Route[]> {
